requirements.txt
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2==2.9.1
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.5
URLObject==2.4.3
Werkzeug==2.0.1
WTForms==2.3.3
